Check value passed to set_header_value_type
This method converts the given symbol to a string, and then compares it
against several literals using ==. This works on JRuby 1.7 but not for
some users on JRuby 9. I'm replacing this with List.contains() and a
constant list of the acceptable values, to both clarify the logic, and
to make the code work on recent JRuby releases.